tensorflow_models/requirements_gpu_1.txt
# GPU-Accelerated BERT Training Requirements - STEP 1
# Install this FIRST: pip install -r requirements_gpu_1.txt
# Optimized for RTX 4070 with TensorFlow DirectML
# Python 3.10.11 compatible

# Core TensorFlow (CPU base - will be accelerated by DirectML plugin in step 2)
tensorflow-cpu==2.10.0
keras>=2.10.0,<2.11.0

# Essential ML Libraries (TF 2.10 Compatible)
numpy>=1.21.0,<1.24.0
scipy>=1.8.0,<1.10.0
pandas>=1.5.0,<2.1.0
scikit-learn>=1.1.0,<1.3.0

# HuggingFace for BERT (TF 2.10 Compatible)
transformers==4.21.0
tokenizers>=0.12.0,<0.14.0
datasets>=2.4.0
accelerate>=0.20.0

# Visualization & Monitoring
matplotlib>=3.5.0,<3.8.0
seaborn>=0.11.0,<0.13.0
tensorboard>=2.10.0,<2.11.0

# Development Tools
tqdm>=4.64.0
jupyter>=1.0.0
ipython>=8.5.0,<8.25.0
notebook>=6.4.0

# Utilities
requests>=2.28.0
pyyaml>=6.0
joblib>=1.2.0

# TensorFlow 2.10 Specific Dependencies
protobuf>=3.19.0,<3.20.0
typing-extensions>=4.2.0,<4.6.0 